+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - feat: declare v0, l1 handler support added
+- fix(fgw): sync from other nodes and block signature
 - fix: added more launcher capabilities
 - fix(cleanup): Updated EditorConfig to 4-space indents
 - fix(tests): Fixed local testing scripts